(add) Zip codes can now be removed from the list, ZipCode component is being mapped instead of hardcoded from state. Improved organization to Redux files.
