+++ src/store/actions/zipCodes.ts
@@ -1,0 +1,12 @@
+export function addZipCode(text: string) {
+  return {
+    type: 'ADD_ZIPCODE',
+    text,
+  };
+}
+export function removeZipCode(text: string) {
+  return {
+    type: 'REMOVE_ZIPCODE',
+    text,
+  };
+}
