--- src/store/zipCodes/actions.ts
@@ -1,6 +1,0 @@
-export function addZipCode(text: string) {
-  return {
-    type: 'ADD_ZIPCODE',
-    text,
-  };
-}
