--- src/store/zipCodes/reducer.ts
@@ -1,14 +1,0 @@
-export default function zipCodes(state = [], action: any) {
-  switch (action.type) {
-    case 'ADD_ZIPCODE':
-      return [
-        ...state,
-        {
-          id: Math.random(),
-          text: action.text,
-        },
-      ];
-    default:
-      return state;
-  }
-}
